Déplacement du code d'import des ecoréseau directement dans la commande correspondante
