+++ scripts/cli.ts
@@ -17,9 +17,9 @@ import { registerJobsCommands } from '@/modules/jobs/commands';
 import { registerOptimizationCommands } from '@/modules/optimization/commands';
 import { registerProEligibilityTestsCommands } from '@/modules/pro-eligibility-tests/commands';
 import { registerNetworkCommands } from '@/modules/reseaux/commands';
+import { registerEcoreseauCommand } from '@/modules/reseaux/commands/ecoreseau';
 import { zAirtableSynchronizableNetworkTable } from '@/modules/reseaux/constants';
 import { downloadNetwork } from '@/modules/reseaux/server/download-network';
-import { importEcoreseauLabels } from '@/modules/reseaux/server/ecoreseaux';
 import { applyGeometryUpdates } from '@/modules/reseaux/server/geometry-updates';
 import { syncPostgresToAirtable } from '@/modules/reseaux/server/sync-pg-to-airtable';
 import { registerTilesCommands } from '@/modules/tiles/commands';
@@ -73,6 +73,7 @@ registerDataCommands(program);
 registerJobsCommands(program);
 registerOptimizationCommands(program);
 registerProEligibilityTestsCommands(program);
+registerEcoreseauCommand(program);
 registerNetworkCommands(program);
 registerOpendataCommands(program);
 registerTilesCommands(program);
@@ -303,13 +304,6 @@ program
       await kdb.updateTable('zones_et_reseaux_en_construction').set({ tags }).where('id_fcu', '=', parseInt(id_fcu_futur, 10)).execute();
     }
     console.info('Tags importés avec succès');
-  });
-
-program
-  .command('import:ecoreseau')
-  .description('Importe les labels Ecoréseau depuis les CSV et met à jour reseaux_de_chaleur.ecoreseau via l’ID SNCU')
-  .action(async () => {
-    await importEcoreseauLabels();
   });
 
 program
